backend/requirements.txt
fastapi
uvicorn[standard]
beanie==2.1.0
lazy-model==0.4.0
motor
mongomock-motor
pydantic
pydantic[email]
pydantic-settings
python-jose[jwt]
passlib[bcrypt]
python-multipart
apscheduler
httpx
openai
pytest
pytest-asyncio
